refactor: relocate internal/lsp/index to internal/index
Pure git-mv of the workspace symbol/edge index out of the LSP
namespace into a peer support package, so cmd/mdsmith can reach
it without importing internal/lsp (the layering map forbids that
edge). Package name unchanged; only import paths and stale path
references in comments/docs updated. internal/schema already
imported it from outside internal/lsp, so the move also removes
a latent layering smell.

Updates the SRP list and DIP arrows in architecture/go.md, the
layering map in architecture/index.md, and records the
supersession of plan 153's symbol-index-stays-LSP-local non-goal
in the architecture audit log.
