+++ internal/index/build_coverage_test.go
