+++ internal/index/complete.go
