+++ internal/index/complete_test.go
