+++ internal/index/coverage_test.go
