+++ internal/index/locate_coverage_test.go
